+++ internal/models/metric/counter.go
@@ -5,14 +5,6 @@ import (
 )
 
 type (
-	// Counter interface {
-	// 	Get() int64
-	// 	Set(v int64) error
-
-	// 	Inc() error
-	// 	Dec() error
-	// }
-
 	Counter int64
 )
