+++ internal/sender/httpsender.go
@@ -16,7 +16,6 @@ type httpSender struct {
 
 func NewSender(serverAddress string) httpSender {
 	baseURL := fmt.Sprintf("http://%s", serverAddress)
-	// client := &http.Client{Timeout: 2 * time.Second}
 	client := &http.Client{}
 
 	return httpSender{baseURL: baseURL, client: client, err: nil}
